Criados métodos dos controllers de Livro e Exemplar, com suas respectivas subclasses

+++ spring-boot-rest-api/src/main/resources/db/migration/V10__Create_Table_Categoria_Livro.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `categoria_livro` (
   `categoria_id` INT NOT NULL,
-  `livro_id` BIGINT NOT NULL,
+  `livro_id` INT NOT NULL,
   PRIMARY KEY (`categoria_id`,`livro_id`),
   KEY `fk_livro_categoria` (`livro_id`),
   CONSTRAINT `fk_categoria_livro` FOREIGN KEY (`categoria_id`) REFERENCES `categoria` (`id`) ON UPDATE CASCADE,
