+++ spring-boot-rest-api/src/main/resources/db/migration/V9__Create_Table_Autor_Livro.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `autor_livro` (
   `autor_id` INT NOT NULL,
-  `livro_id` BIGINT NOT NULL,
+  `livro_id` INT NOT NULL,
   PRIMARY KEY (`autor_id`,`livro_id`),
   KEY `fk_livro_autor` (`livro_id`),
   CONSTRAINT `fk_autor_livro` FOREIGN KEY (`autor_id`) REFERENCES `autor` (`id`) ON UPDATE CASCADE,
